LPS-196245 Delete DDLWebConfigurationActivator and move configuration to its usages
