--- modules/apps/dynamic-data-lists/dynamic-data-lists-web/src/main/java/com/liferay/dynamic/data/lists/web/internal/configuration/activator/DDLWebConfigurationActivator.java
@@ -1,39 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.dynamic.data.lists.web.internal.configuration.activator;
-
-import com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfiguration;
-import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-
-import java.util.Map;
-
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Modified;
-
-/**
- * @author Leonardo Barros
- */
-@Component(
-	configurationPid = "com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfiguration",
-	service = DDLWebConfigurationActivator.class
-)
-public class DDLWebConfigurationActivator {
-
-	public DDLWebConfiguration getDDLWebConfiguration() {
-		return _ddlWebConfiguration;
-	}
-
-	@Activate
-	@Modified
-	protected void activate(Map<String, Object> properties) {
-		_ddlWebConfiguration = ConfigurableUtil.createConfigurable(
-			DDLWebConfiguration.class, properties);
-	}
-
-	private volatile DDLWebConfiguration _ddlWebConfiguration;
-
-}
