+++ modules/apps/dynamic-data-lists/dynamic-data-lists-web/src/main/java/com/liferay/dynamic/data/lists/web/internal/portlet/DDLDisplayPortlet.java
@@ -14,12 +14,13 @@ import com.liferay.dynamic.data.lists.service.DDLRecordService;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetLocalService;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetService;
 import com.liferay.dynamic.data.lists.util.DDL;
-import com.liferay.dynamic.data.lists.web.internal.configuration.activator.DDLWebConfigurationActivator;
+import com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfiguration;
 import com.liferay.dynamic.data.lists.web.internal.display.context.DDLDisplayContext;
 import com.liferay.dynamic.data.mapping.security.permission.DDMPermissionSupport;
 import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
 import com.liferay.dynamic.data.mapping.storage.DDMStorageEngineManager;
 import com.liferay.fragment.processor.PortletRegistry;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.PortletPreferencesException;
 import com.liferay.portal.kernel.log.Log;
@@ -35,6 +36,8 @@ import com.liferay.portal.kernel.util.WebKeys;
 
 import java.io.IOException;
 
+import java.util.Map;
+
 import javax.portlet.Portlet;
 import javax.portlet.PortletException;
 import javax.portlet.RenderRequest;
@@ -43,12 +46,14 @@ import javax.portlet.RenderResponse;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Modified;
 import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Marcellus Tavares
  */
 @Component(
+	configurationPid = "com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfiguration",
 	property = {
 		"com.liferay.portlet.add-default-resource=true",
 		"com.liferay.portlet.css-class-wrapper=portlet-dynamic-data-lists-display",
@@ -90,9 +95,8 @@ public class DDLDisplayPortlet extends MVCPortlet {
 
 			DDLDisplayContext ddlDisplayContext = new DDLDisplayContext(
 				renderRequest, renderResponse, _ddl, _ddlRecordSetLocalService,
-				_ddlWebConfigurationActivator.getDDLWebConfiguration(),
-				_ddmPermissionSupport, _ddmStorageEngineManager,
-				_ddmTemplateLocalService);
+				_ddlWebConfiguration, _ddmPermissionSupport,
+				_ddmStorageEngineManager, _ddmTemplateLocalService);
 
 			renderRequest.setAttribute(
 				WebKeys.PORTLET_DISPLAY_CONTEXT, ddlDisplayContext);
@@ -116,7 +120,9 @@ public class DDLDisplayPortlet extends MVCPortlet {
 	}
 
 	@Activate
-	protected void activate() {
+	protected void activate(Map<String, Object> properties) {
+		modified(properties);
+
 		_portletRegistry.registerAlias(
 			_ALIAS, DDLPortletKeys.DYNAMIC_DATA_LISTS_DISPLAY);
 	}
@@ -156,6 +162,12 @@ public class DDLDisplayPortlet extends MVCPortlet {
 		}
 
 		return false;
+	}
+
+	@Modified
+	protected void modified(Map<String, Object> properties) {
+		_ddlWebConfiguration = ConfigurableUtil.createConfigurable(
+			DDLWebConfiguration.class, properties);
 	}
 
 	protected void setDDLRecordRequestAttribute(RenderRequest renderRequest)
@@ -206,8 +218,7 @@ public class DDLDisplayPortlet extends MVCPortlet {
 	@Reference
 	private DDLRecordSetService _ddlRecordSetService;
 
-	@Reference
-	private DDLWebConfigurationActivator _ddlWebConfigurationActivator;
+	private volatile DDLWebConfiguration _ddlWebConfiguration;
 
 	@Reference
 	private DDMPermissionSupport _ddmPermissionSupport;
