+++ modules/apps/dynamic-data-lists/dynamic-data-lists-web/src/main/java/com/liferay/dynamic/data/lists/web/internal/portlet/action/ExportRecordSetMVCResourceCommand.java
@@ -12,8 +12,8 @@ import com.liferay.dynamic.data.lists.model.DDLRecordSet;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetService;
 import com.liferay.dynamic.data.lists.util.comparator.DDLRecordModifiedDateComparator;
 import com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfiguration;
-import com.liferay.dynamic.data.lists.web.internal.configuration.activator.DDLWebConfigurationActivator;
 import com.liferay.petra.string.CharPool;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.portlet.PortletResponseUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCResourceCommand;
@@ -25,16 +25,21 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
+import java.util.Map;
+
 import javax.portlet.ResourceRequest;
 import javax.portlet.ResourceResponse;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Modified;
 import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Marcellus Tavares
  */
 @Component(
+	configurationPid = "com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfiguration",
 	property = {
 		"javax.portlet.name=" + DDLPortletKeys.DYNAMIC_DATA_LISTS,
 		"javax.portlet.name=" + DDLPortletKeys.DYNAMIC_DATA_LISTS_DISPLAY,
@@ -44,6 +49,13 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class ExportRecordSetMVCResourceCommand extends BaseMVCResourceCommand {
 
+	@Activate
+	@Modified
+	protected void activate(Map<String, Object> properties) {
+		_ddlWebConfiguration = ConfigurableUtil.createConfigurable(
+			DDLWebConfiguration.class, properties);
+	}
+
 	@Override
 	protected void doServeResource(
 			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
@@ -52,11 +64,8 @@ public class ExportRecordSetMVCResourceCommand extends BaseMVCResourceCommand {
 		String fileExtension = ParamUtil.getString(
 			resourceRequest, "fileExtension");
 
-		DDLWebConfiguration ddlWebConfiguration =
-			_ddlWebConfigurationActivator.getDDLWebConfiguration();
-
 		if (StringUtil.equals(fileExtension, "csv") &&
-			StringUtil.equals(ddlWebConfiguration.csvExport(), "disabled")) {
+			StringUtil.equals(_ddlWebConfiguration.csvExport(), "disabled")) {
 
 			return;
 		}
@@ -92,7 +101,6 @@ public class ExportRecordSetMVCResourceCommand extends BaseMVCResourceCommand {
 	@Reference
 	private DDLRecordSetService _ddlRecordSetService;
 
-	@Reference
-	private DDLWebConfigurationActivator _ddlWebConfigurationActivator;
+	private volatile DDLWebConfiguration _ddlWebConfiguration;
 
 }
